Fix tipo variabile da int a long per id prodotto

+++ src/Prodotti/ProdottiService/Models/API/ProductDTO.cs
@@ -2,13 +2,13 @@
 {
     public class ProductDTO
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Nome { get; set; }
         public string Descrizione { get; set; }
         public double Prezzo { get; set; }
         public int QuantitaDisponibile { get; set; }
 
-        public ProductDTO(int id, string nome, string descrizione, double prezzo, int quantitaDisponibile)
+        public ProductDTO(long id, string nome, string descrizione, double prezzo, int quantitaDisponibile)
         {
             this.Id = id;
             this.Nome = nome;
@@ -17,7 +17,7 @@
             this.QuantitaDisponibile = quantitaDisponibile;
         }
 
-        public static ProductDTO ProductDTOFactory(int id, string nome, string descrizione, double prezzo, int quantitaDisponibile)
+        public static ProductDTO ProductDTOFactory(long id, string nome, string descrizione, double prezzo, int quantitaDisponibile)
         {
             return new ProductDTO(id, nome, descrizione, prezzo, quantitaDisponibile);
         }
